chore: rm widget components
