--- src/lib/components/Swap/Settings.fixture.tsx
@@ -1,8 +1,0 @@
-import { Modal } from '../Dialog'
-import { SettingsDialog } from './Settings'
-
-export default (
-  <Modal color="module">
-    <SettingsDialog />
-  </Modal>
-)
