--- src/lib/components/Swap/Status.fixture.tsx
@@ -1,12 +1,0 @@
-import { Modal } from '../Dialog'
-
-function Fixture() {
-  return null
-  // TODO(zzmp): Mock <StatusDialog tx={} onClose={() => void 0} />
-}
-
-export default (
-  <Modal color="dialog">
-    <Fixture />
-  </Modal>
-)
