--- src/lib/components/Swap/Status/index.ts
@@ -1,1 +1,0 @@
-export { default as StatusDialog } from './StatusDialog'
